Avoid unsigned overflow in length computations
Additions like

    hdr->iv_length + hdr->auth_tag_length + hdr->kdf_context_length

and

    length += sizeof(hdr->auth_tag_length) + hdr->auth_tag_length

can cause unsigned overflow during itermediate computations. This can
lead to incorrect buffer lengths being computed, causing further pointer
computations to use invalid memory addresses, resuling in arbitrary
memory writes controlled by an attacker -- since these lengths come
from message data structure that we parse.

This is further complicated by the fact that "size_t" is typically
32-bit on 32-bit systems, causing more possible oveflows.

Bad news here is that unsigned integer overflow is *not* undefined
behavior in C -- it is defined to be two's complement. Therefore most
tools like UBSan and static analyzers don't catch issues due to them
not being "undefined behavior", so they are not reported. This issue
has been caught only due to its 'side effect' in parsing.

The only way to avoid issues is to either check for overflows or avoid
them. Both options require careful coding of each arithmetic operation,
so we opt for avoiding overflows without significant runtime cost. We
update the code to use "uint64_t" where the maximum size can exceed
uint32_t range (we know that it won't overflow uint64_t because there
are only a handful of values added), and similarly use "uint32_t" for
cases where the value cannot overflow "uint32_t".

Hopefully this is it... By the way, this is *the* reason why Rust
does not have implicit casts between integer types.

+++ src/themis/secure_cell_seal_passphrase.c
@@ -94,7 +94,7 @@ static themis_status_t themis_auth_sym_derive_encryption_key_pbkdf2(
 
     /* KDF pointer is ignored but size is important */
     hdr->kdf_context = NULL;
-    hdr->kdf_context_length = (uint32_t)themis_scell_pbkdf2_context_size(&kdf);
+    hdr->kdf_context_length = themis_scell_pbkdf2_context_size(&kdf);
 
     res = themis_write_scell_pbkdf2_context(hdr, &kdf, auth_token, *auth_token_length);
     if (res != THEMIS_SUCCESS) {
